requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pillow==12.1.1
pillow-heif==1.3.0
aiofiles==24.1.0
